Fixes the intent of my tool spell
I made this spell for use in Solaris back when that was a thing, they ported it over to AP but changed it to being miracle-exclusive. That was never the intent, so I've added it back to the mage list and added some more tools to the selection as well!