No, it doesn't slow down by 50% anymore
More clarification. At the very least if this doesn't get put in, we can go back and change the note so it only says 50% (which was true before) or add a clarification where instead of saying it say to the equivalent of a speed 5 character, instead to something like a theoretical speed -5 character.